raph-claude-attempt/requirements.txt
pyrplidar>=0.1.2
matplotlib>=3.5
numpy>=1.21
pyyaml>=6.0
